refactor(shared): states / parsePackageType を packageDisplay に分離
メインワールドの React から import する経路に fs-extra(shared/install
経由)が混入すると renderer の webpack ビルドが失敗するため、fs 非依存の
表示用モジュールとして切り出す。既存の import 経路は再輸出で維持。
